+++ pyrus/models/responses.py
@@ -5,6 +5,7 @@
 from . import entities
 from . import constants
 
+
 class BaseResponse(object):
     """
         error_code (:obj:`str`): Error code
@@ -21,7 +22,8 @@ class BaseResponse(object):
             self.error_code = kwargs['error_code']
         if 'error' in kwargs:
             self.error = kwargs['error']
-        
+
+
 class AuthResponse(BaseResponse):
     """
         AuthResponse
@@ -42,6 +44,7 @@ class AuthResponse(BaseResponse):
             self.success = False
         super(AuthResponse, self).__init__(**kwargs)
 
+
 class FormResponse(BaseResponse):
     """
         FormResponse
@@ -49,10 +52,12 @@ class FormResponse(BaseResponse):
         Attributes:
             id (:obj:`int`): Form id
             name (:obj:`str`): Form name
-            steps (:obj:`dict`): Form name
+            steps (:obj:`dict`): Form steps
                 key (:obj:`int`): Step number
                 value (:obj:`str`): Step name
             fields (:obj:`list` of :obj:`models.entities.FormField`): List of form fields
+            deletedOrClosed (:obj:`bool`): Form state
+            folder (:obj:`list` of :obj:`str`): Folder of form
     """
     __doc__ += BaseResponse.__doc__
 
@@ -60,6 +65,9 @@ class FormResponse(BaseResponse):
     name = None
     steps = None
     fields = None
+    deletedOrClosed = None
+    folder = None
+
     @property
     def flat_fields(self):
         return self._get_flat_fields(self.fields)
@@ -75,6 +83,12 @@ class FormResponse(BaseResponse):
             self.fields = []
             for field in kwargs['fields']:
                 self.fields.append(entities.FormField(**field))
+        if 'deleted_or_closed' in kwargs:
+            self.deletedOrClosed = kwargs['deleted_or_closed']
+        if 'folder' in kwargs:
+            self.folder = []
+            for fld in kwargs['folder']:
+                self.folder.append(fld)
         super(FormResponse, self).__init__(**kwargs)
 
     def _get_flat_fields(self, fields):
@@ -92,9 +106,9 @@ class FormResponse(BaseResponse):
                     res.extend(self._get_flat_fields(option.fields))
             elif field.info.columns:
                 res.extend(field.info.columns)
-            
+
         return res
-        
+
 
 class FormsResponse(BaseResponse):
     """
@@ -114,6 +128,7 @@ class FormsResponse(BaseResponse):
                 self.forms.append(FormResponse(**form))
         super(FormsResponse, self).__init__(**kwargs)
 
+
 class TaskResponse(BaseResponse):
     """
         TaskResponse
@@ -124,11 +139,12 @@ class TaskResponse(BaseResponse):
     __doc__ += BaseResponse.__doc__
 
     task = None
-    
+
     def __init__(self, **kwargs):
         if 'task' in kwargs:
             self.task = entities.TaskWithComments(**kwargs['task'])
         super(TaskResponse, self).__init__(**kwargs)
+
 
 class ContactsResponse(BaseResponse):
     """
@@ -147,6 +163,7 @@ class ContactsResponse(BaseResponse):
             for organization in kwargs['organizations']:
                 self.organizations.append(entities.Organization(**organization))
         super(ContactsResponse, self).__init__(**kwargs)
+
 
 class CatalogResponse(BaseResponse):
     """
@@ -176,6 +193,7 @@ class CatalogResponse(BaseResponse):
                 self.catalog_headers.append(entities.CatalogHeader(**item))
         super(CatalogResponse, self).__init__(**kwargs)
 
+
 class FormRegisterResponse(BaseResponse):
     """
         FormRegisterResponse
@@ -196,6 +214,7 @@ class FormRegisterResponse(BaseResponse):
                 self.tasks.append(entities.Task(**task))
         super(FormRegisterResponse, self).__init__(**kwargs)
 
+
 class UploadResponse(BaseResponse):
     """
         UploadResponse
@@ -208,13 +227,14 @@ class UploadResponse(BaseResponse):
 
     guid = None
     md5_hash = None
-    
+
     def __init__(self, **kwargs):
         if 'guid' in kwargs:
             self.guid = kwargs['guid']
         if 'md5_hash' in kwargs:
             self.md5_hash = kwargs['md5_hash']
         super(UploadResponse, self).__init__(**kwargs)
+
 
 class ListsResponse(BaseResponse):
     """
@@ -234,6 +254,7 @@ class ListsResponse(BaseResponse):
                 self.lists.append(entities.TaskList(**lst))
         super(ListsResponse, self).__init__(**kwargs)
 
+
 class TaskListResponse(BaseResponse):
     """
         TaskListResponse
@@ -246,7 +267,7 @@ class TaskListResponse(BaseResponse):
 
     tasks = None
     has_more = None
-    
+
     def __init__(self, **kwargs):
         if 'has_more' in kwargs:
             self.has_more = kwargs['has_more']
@@ -255,6 +276,7 @@ class TaskListResponse(BaseResponse):
             for task in kwargs['tasks']:
                 self.tasks.append(entities.TaskHeader(**task))
         super(TaskListResponse, self).__init__(**kwargs)
+
 
 class DownloadResponse(BaseResponse):
     """
@@ -274,6 +296,7 @@ class DownloadResponse(BaseResponse):
         self.raw_file = raw_file
         super(DownloadResponse, self).__init__(**{})
 
+
 class SyncCatalogResponse(BaseResponse):
     """
         SyncCatalogResponse
@@ -292,7 +315,7 @@ class SyncCatalogResponse(BaseResponse):
     deleted = None
     updated = None
     catalog_headers = None
-    
+
     def __init__(self, **kwargs):
         if 'apply' in kwargs:
             self.apply = kwargs['apply']
@@ -313,3 +336,43 @@ class SyncCatalogResponse(BaseResponse):
             for header in kwargs['catalog_headers']:
                 self.catalog_headers.append(entities.CatalogHeader(**header))
         super(SyncCatalogResponse, self).__init__(**kwargs)
+
+class ProfileResponse(BaseResponse):
+    """
+        ProfileResponse
+        
+        Attributes:
+            person_id (:obj:`int`): Person id
+            first_name (:obj:`str`): Person first name
+            last_name (:obj:`str`): Person last name
+            email(:obj:`str`): Person email
+            locale(:obj:`str`): Person locale (ru-RU/en-US/en-GB)
+            organization_id(:obj:`int`) Person organization id
+            organization(:obj:`obj`models.entities.Organization) Person organization
+    """
+    __doc__ += BaseResponse.__doc__
+
+    person_id = None
+    first_name = None
+    last_name = None
+    email = None
+    locale = None
+    organization_id = None
+    organization = None
+
+    def __init__(self, **kwargs):
+        if 'person_id' in kwargs:
+            self.person_id = kwargs['person_id']
+        if 'first_name' in kwargs:
+            self.first_name = kwargs['first_name']
+        if 'last_name' in kwargs:
+            self.last_name = kwargs['last_name']
+        if 'email' in kwargs:
+            self.email = kwargs['email']
+        if 'locale' in kwargs:
+            self.locale = kwargs['locale']
+        if 'organization_id' in kwargs:
+            self.organization_id = kwargs['organization_id']
+        if 'organization' in kwargs:
+            self.organization = entities.Organization(**kwargs['organization'])
+        super(ProfileResponse, self).__init__(**kwargs)